requirements.txt
typer
rich
docker
openai
pyyaml
jinja2
prompt_toolkit
gnureadline; sys_platform == 'darwin'
chromadb
pydantic-settings
